# lookup4 (new UTF-8 validation)