Motivated by a tweet by Z. Wegner who pointed out that lookup3 was probably not optimal. I think he is correct.

This lookup4 is faster.