@michaeleisel 2020-06-28
fyi here's rust utf-8 validation: https://github.com/bluss/rust/blob/cadcd70/src/libcore/str/mod.rs#L1095
swift utf-8 validation is not open-source (see `_tryFromUTF8` in the swift source code), although i can get the assembly if it'd help