@lemire 2020-06-28
I bet we can get the swift team to adopt lookup4.