@jkeiser 2020-06-29
Overall numbers (skylake):

| File                                | master Cycles | PR Cycles | +Throughput | master Instr. | PR Instr. | -Instr. | master Branch Misses | PR Branch Misses | -Branch Misses | master Cache Misses | PR Cache Misses | -Cache Misses | master Cache Refs | PR Cache Refs | -Cache Refs |
| ---                                 |   --: |   --: |   --: |   --: |   --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |
| repeat                              | 102.5 | 100.2 |    2% | 354.5 | 346.9 |    2% |     12 |     13 |   -8% |      0 |      0 |    0% |      7 |     14 | -100% |
| random                              | 138.3 | 136.6 |    1% | 480.4 | 474.1 |    1% |   1974 |   1964 |    0% |      0 |      0 |    0% |  86744 |  86675 |    0% |
| twitter                             |  92.0 |  91.2 |    0% | 291.3 | 288.7 |    0% |   3480 |   3400 |    2% |      0 |      0 |    0% |  78643 |  78639 |    0% |
| twitter_timeline                    | 129.9 | 128.9 |    0% | 396.5 | 396.0 |    0% |    125 |    133 |   -6% |      0 |      0 |    0% |   2244 |   1088 |   51% |
| instruments                         |  95.1 |  94.4 |    0% | 340.0 | 339.5 |    0% |    512 |    404 |   21% |      0 |      0 |    0% |  29628 |  29996 |   -1% |
| tree-pretty                         |  94.7 |  94.3 |    0% | 332.2 | 331.7 |    0% |     54 |     51 |    5% |      0 |      0 |    0% |    821 |    690 |   15% |
| twitter_api_response                |  89.9 |  89.8 |    0% | 302.4 | 301.5 |    0% |     18 |     19 |   -5% |      0 |      0 |    0% |     98 |     44 |   55% |
| apache_builds                       |  87.5 |  87.4 |    0% | 307.8 | 307.3 |    0% |    116 |     96 |   17% |      0 |      0 |    0% |  13840 |  13963 |    0% |
| google_maps_api_compact_response    | 193.2 | 193.0 |    0% | 674.5 | 673.7 |    0% |     64 |     64 |    0% |      0 |      0 |    0% |    120 |    108 |   10% |
| twitterescaped                      | 192.0 | 192.0 |    0% | 554.0 | 553.5 |    0% |   5075 |   4966 |    2% |      0 |      0 |    0% |  75144 |  75100 |    0% |
| google_maps_api_response            | 103.7 | 103.7 |    0% | 356.8 | 356.3 |    0% |     62 |     65 |   -4% |      0 |      0 |    0% |   1082 |    348 |   67% |
| mesh.pretty                         | 166.0 | 166.0 |    0% | 512.1 | 511.6 |    0% |  13339 |  13475 |   -1% |   1936 |      2 |   99% | 181085 | 180823 |    0% |
| update-center                       | 113.5 | 113.6 |    0% | 338.8 | 338.2 |    0% |   5986 |   5904 |    1% |      0 |      0 |    0% |  79479 |  79428 |    0% |
| numbers                             | 229.0 | 229.1 |    0% | 704.1 | 703.6 |    0% |   2168 |   2168 |    0% |      0 |      0 |    0% |  17921 |  18616 |   -3% |
| github_events                       |  80.0 |  80.1 |    0% | 266.0 | 265.5 |    0% |    110 |    112 |   -1% |      0 |      0 |    0% |   3641 |   3345 |    8% |
| mesh                                | 269.5 | 269.8 |    0% | 841.5 | 840.9 |    0% |   9846 |   9896 |    0% |      0 |      0 |    0% | 127663 | 127911 |    0% |
| canada                              | 275.0 | 275.5 |    0% | 894.5 | 894.0 |    0% |  18706 |  18731 |    0% |  37395 |  40621 |   -8% | 323308 | 323589 |    0% |
| marine_ik                           | 263.1 | 263.7 |    0% | 792.2 | 791.6 |    0% |  48785 |  48960 |    0% | 260424 | 275546 |   -5% | 556020 | 556261 |    0% |
| twitter_api_compact_response        | 111.3 | 111.6 |    0% | 371.3 | 370.2 |    0% |     16 |     15 |    6% |      0 |      0 |    0% |     48 |     71 |  -47% |
| citm_catalog                        |  79.3 |  79.6 |    0% | 293.5 | 292.9 |    0% |   1735 |   1751 |    0% |      0 |   2428 |    0% | 182374 | 182603 |    0% |
| gsoc-2018                           |  70.9 |  72.0 |   -1% | 170.5 | 170.0 |    0% |  30021 |  29983 |    0% |  91004 | 103893 |  -14% | 350907 | 351785 |    0% |

Stage 1 only (skylake):

| File                                | master Cycles | PR Cycles | +Throughput | master Instr. | PR Instr. | -Instr. | master Branch Misses | PR Branch Misses | -Branch Misses | master Cache Misses | PR Cache Misses | -Cache Misses | master Cache Refs | PR Cache Refs | -Cache Refs |
| ---                                 |   --: |   --: |   --: |   --: |   --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |
| repeat                              |  50.8 |  48.5 |    4% | 187.8 | 180.2 |    4% |      1 |      1 |    0% |      0 |      0 |    0% |     14 |      2 |   85% |
| random                              |  57.8 |  55.9 |    3% | 209.3 | 203.0 |    3% |   1037 |    995 |    4% |      0 |      0 |    0% |  27743 |  27726 |    0% |
| instruments                         |  38.3 |  37.7 |    1% | 143.1 | 142.6 |    0% |    175 |     85 |   51% |      0 |      0 |    0% |  10282 |  10421 |   -1% |
| twitter                             |  43.7 |  43.1 |    1% | 148.8 | 146.2 |    1% |   1795 |   1770 |    1% |      0 |      0 |    0% |  27164 |  27173 |    0% |
| google_maps_api_response            |  36.3 |  36.0 |    0% | 137.2 | 136.7 |    0% |      1 |      1 |    0% |      0 |      0 |    0% |    483 |    102 |   78% |
| twitter_api_compact_response        |  41.5 |  41.2 |    0% | 153.1 | 152.0 |    0% |      1 |      1 |    0% |      0 |      0 |    0% |     17 |     12 |   29% |
| apache_builds                       |  36.7 |  36.5 |    0% | 139.3 | 138.8 |    0% |     24 |      6 |   75% |      0 |      0 |    0% |   5103 |   5228 |   -2% |
| update-center                       |  44.9 |  44.6 |    0% | 142.1 | 141.6 |    0% |   2861 |   2722 |    4% |      0 |      0 |    0% |  25253 |  25299 |    0% |
| twitterescaped                      |  41.8 |  41.6 |    0% | 138.1 | 137.5 |    0% |   1938 |   1833 |    5% |      0 |      0 |    0% |  25118 |  25095 |    0% |
| twitter_api_response                |  37.8 |  37.7 |    0% | 140.1 | 139.3 |    0% |      1 |      1 |    0% |      0 |      0 |    0% |     31 |     10 |   67% |
| mesh                                |  46.1 |  46.0 |    0% | 174.2 | 173.5 |    0% |    772 |    752 |    2% |      0 |      0 |    0% |  43035 |  43023 |    0% |
| tree-pretty                         |  36.7 |  36.7 |    0% | 140.6 | 140.1 |    0% |      2 |      1 |   50% |      0 |      0 |    0% |    254 |    247 |    2% |
| mesh.pretty                         |  35.7 |  35.6 |    0% | 137.7 | 137.2 |    0% |     35 |     41 |  -17% |    105 |      3 |   97% |  69433 |  69390 |    0% |
| twitter_timeline                    |  40.3 |  40.3 |    0% | 153.5 | 153.0 |    0% |      2 |      5 | -150% |      0 |      0 |    0% |    701 |    314 |   55% |
| google_maps_api_compact_response    |  49.8 |  49.9 |    0% | 190.3 | 189.5 |    0% |      2 |      1 |   50% |      0 |      0 |    0% |     66 |     35 |   46% |
| github_events                       |  34.2 |  34.2 |    0% | 129.8 | 129.3 |    0% |      4 |      6 |  -50% |      0 |      0 |    0% |   1550 |   1514 |    2% |
| citm_catalog                        |  34.0 |  34.1 |    0% | 130.7 | 130.2 |    0% |    166 |    174 |   -4% |      0 |    576 |    0% |  71650 |  71882 |    0% |
| marine_ik                           |  53.8 |  54.1 |    0% | 180.9 | 180.4 |    0% |   8678 |   8729 |    0% |  68490 |  72094 |   -5% | 186005 | 186213 |    0% |
| canada                              |  44.7 |  45.0 |    0% | 157.6 | 157.1 |    0% |   5266 |   5335 |   -1% |   5472 |   9792 |  -78% | 114958 | 115417 |    0% |
| gsoc-2018                           |  36.3 |  36.6 |    0% | 101.1 | 100.6 |    0% |  14812 |  15010 |   -1% |  21048 |  27417 |  -30% | 116300 | 116947 |    0% |
| numbers                             |  38.6 |  39.0 |   -1% | 144.3 | 143.8 |    0% |    112 |    148 |  -32% |      0 |      0 |    0% |   6983 |   7016 |    0% |